\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,ExponentialData-method}
\alias{show,ExponentialLuckModel-method}
\alias{show,LuckModel-method}
\alias{show,LuckModelData-method}
\alias{show,ScaledNormalData-method}
\alias{show,ScaledNormalLuckModel-method}
\title{ Methods for printing LuckModel and LuckModelData objects of package `luck' }
\description{
The output at the prompt for objects in the S4 system is generated by the function \code{show},
as compared to the function \code{print} for objects in the S3 system.
If a \code{\linkS4class{LuckModel}} object contains a \code{\linkS4class{LuckModelData}}
object, the \code{show} output of the \code{\linkS4class{LuckModelData}} object
is included in the \code{show} output for the \code{\linkS4class{LuckModel}} object,
i.e., information on the data contained in the \code{\linkS4class{LuckModel}} object
is displayed as well.
For plain \code{\linkS4class{LuckModel}} objects, just the bounds for the
canonical parameters of the prior are displayed, whereas for subclasses of \code{\linkS4class{LuckModel}}
objects also the translation to the usual prior parameters are shown.
}
\seealso{
\code{\link{luck}} for a general description of the package,
\code{\linkS4class{LuckModel}} as the central class of the package,
\code{\linkS4class{LuckModelData}} as the class for its \code{data} slot,
\code{\linkS4class{ScaledNormalLuckModel}} and
\code{\linkS4class{ExponentialLuckModel}} for non-abstract subclasses intended
for analysis of scaled normal or exponential data, respectively.
}
\examples{
# generate a generic generalized iLUCK model without data
luck1 <- LuckModel(n0=c(2,10), y0=c(3, 4))
# display the model (this invokes show())
luck1
show(luck1) # same result
# add a LuckModelData object to the LuckModel object
data(luck1) <- LuckModelData(tau=20, n=10)
# display the model
luck1
# generate a generalized iLUCK model for inference from exponentially distributed data
expluck <- ExponentialLuckModel(y0=c(1,3), n0=c(1,2), data=ExponentialData(mean=mean(rexp(5)), n=5))
# display the model
expluck
}
\keyword{methods}
\keyword{print}
